requirements.txt
flask==2.0.1
pytesseract==0.3.13
opencv-python==4.5.3.56
numpy==1.21.2
easyocr==1.4.1
scikit-image==0.22.0
Pillow==8.3.2
